feat(auth): split up authentication into its own component
