+++ src/app/modules/authentication/components/authentication/authentication.component.scss
@@ -1,0 +1,91 @@
+@use "sass:math";
+@import "variables";
+
+$container-height: 400px;
+
+.margin-button {
+	margin-top: 8px;
+}
+
+.login-containers {
+	display: grid;
+	grid-template-columns: 1fr 1fr;
+	grid-auto-flow: row;
+
+	column-gap: 30px;
+
+	.login-container {
+		background-color: $tile-background-color;
+		overflow: hidden;
+
+		height: 100%;
+
+		.login-icon-container {
+			display: flex;
+
+			align-items: center;
+			justify-content: center;
+
+			height: math.div($container-height, 2);
+
+			background: rgb(64, 49, 188);
+			background: linear-gradient(0deg, rgba(64, 49, 188, 1) 0%, rgba(49, 129, 229, 1) 100%);
+
+			.icon {
+				width: 150px;
+				height: 150px;
+				line-height: 150px;
+				border-radius: 50%;
+
+				&.osu {
+					background-image: url("~assets/images/osulogo.png");
+					background-size: cover;
+					background-repeat: no-repeat;
+				}
+
+				&.web {
+					background-image: url("~assets/images/icon.png");
+					background-size: cover;
+					background-repeat: no-repeat;
+				}
+			}
+		}
+
+		.login-content {
+			margin: 38px;
+
+			.login-form {
+				background-color: darken($tile-background-color, 5%);
+
+				padding: 8px;
+
+				-webkit-box-shadow: 2px 3px 5px 0px rgba(0, 0, 0, 0.75);
+				-moz-box-shadow: 2px 3px 5px 0px rgba(0, 0, 0, 0.75);
+				box-shadow: 2px 3px 5px 0px rgba(0, 0, 0, 0.75);
+
+				.header {
+					text-align: center;
+				}
+
+				.center {
+					display: flex;
+					justify-content: center;
+
+					margin-top: 8px;
+
+					.text {
+						padding-left: 8px;
+						display: flex;
+						align-self: center;
+					}
+				}
+
+				.footer {
+					margin-top: 8px;
+
+					text-align: center;
+				}
+			}
+		}
+	}
+}
