+++ src/app/modules/settings/components/settings/settings.component.ts
@@ -3,7 +3,6 @@ import { ElectronService } from '../../../../services/electron.service';
 import { StoreService } from '../../../../services/store.service';
 import { ToastService } from '../../../../services/toast.service';
 import { ToastType } from '../../../../models/toast';
-import { ApiKeyValidation } from '../../../../services/osu-api/api-key-validation.service';
 import { MatDialog } from '@angular/material/dialog';
 import { RemoveSettingsComponent } from '../../../../components/dialogs/remove-settings/remove-settings.component';
 import { AuthenticateService } from 'app/services/authenticate.service';
@@ -18,22 +17,10 @@ import { GenericService } from 'app/services/generic.service';
 })
 
 export class SettingsComponent implements OnInit {
-	apiKey: string;
-	isAuthenticating: boolean;
-
 	dialogMessage: string;
 	dialogAction = 0;
 
-	mappoolPublishForm: FormGroup;
-	ircLoginForm: FormGroup;
-
-	isConnecting = false;
-	isDisconnecting = false;
-
-	apiKeyIsValid = false;
-
 	axsMenuStatus: boolean;
-
 	splitIrcMessages: boolean;
 
 	allOptions: { icon: string; message: string; buttonText: string; action: any }[] = [
@@ -47,117 +34,17 @@ export class SettingsComponent implements OnInit {
 		public electronService: ElectronService,
 		private storeService: StoreService,
 		private toastService: ToastService,
-		private apiKeyValidation: ApiKeyValidation,
 		private dialog: MatDialog,
 		public authService: AuthenticateService,
 		public ircService: IrcService,
 		private genericService: GenericService
 	) {
-		this.apiKey = this.storeService.get('api-key');
-		this.isAuthenticating = false;
-
-		if (this.apiKey && this.apiKey.length > 0) {
-			this.apiKeyIsValid = true;
-		}
-
 		this.genericService.getAxSMenuStatus().subscribe(status => {
 			this.axsMenuStatus = status;
 		});
 	}
 
-	ngOnInit() {
-		this.mappoolPublishForm = new FormGroup({
-			username: new FormControl('', [
-				Validators.required
-			]),
-			password: new FormControl('', [
-				Validators.required
-			])
-		});
-
-		this.ircLoginForm = new FormGroup({
-			'irc-username': new FormControl('', [
-				Validators.required
-			]),
-			'irc-password': new FormControl('', [
-				Validators.required
-			])
-		});
-
-		// Subscribe to the isConnecting variable to show/hide the spinner
-		this.ircService.getIsConnecting().subscribe(value => {
-			this.isConnecting = value;
-		});
-
-		// Subscribe to the isConnecting variable to show/hide the spinner
-		this.ircService.getIsDisconnecting().subscribe(value => {
-			this.isDisconnecting = value;
-		});
-	}
-
-	/**
-	 * Start osu authentication process
-	 */
-	authenticateOsu(): void {
-		this.isAuthenticating = true;
-
-		this.authService.startOsuOauthProcess().subscribe(token => {
-			if (token != null) {
-				this.authService.handleOauth(token).subscribe(user => {
-					this.authService.loginUser(user);
-				});
-			}
-
-			this.isAuthenticating = false;
-		});
-	}
-
-	logoutOsu(): void {
-		this.authService.logout();
-	}
-
-	/**
-	 * Login to irc with the given credentials
-	 */
-	connectIrc() {
-		const username = this.ircLoginForm.get('irc-username').value;
-		const password = this.ircLoginForm.get('irc-password').value;
-
-		this.ircService.connect(username, password);
-	}
-
-	disconnectIrc() {
-		this.ircService.disconnect();
-	}
-
-	/**
-	 * Get the api key
-	 */
-	getApiKey() {
-		return this.storeService.get('api-key');
-	}
-
-	/**
-	 * Save the api key with the entered value
-	 */
-	saveApiKey() {
-		// Key is valid
-		this.apiKeyValidation.validate(this.apiKey).subscribe(() => {
-			this.storeService.set('api-key', this.apiKey);
-			this.toastService.addToast('You have entered a valid api-key.', ToastType.Information);
-			this.toastService.addToast('The client will reload itself in 10 seconds.', ToastType.Information, 10);
-
-			this.apiKeyIsValid = true;
-
-			setTimeout(() => {
-				window.location.reload();
-			}, 10000);
-		},
-			// Key is invalid
-			() => {
-				this.toastService.addToast('The entered api-key was invalid.', ToastType.Error);
-			});
-	}
+	ngOnInit() { }
 
 	/**
 	 * Clear the cache
@@ -173,9 +60,6 @@ export class SettingsComponent implements OnInit {
 	removeApiKey() {
 		this.storeService.delete('api-key');
 		this.toastService.addToast('Successfully removed your api key.');
-
-		this.apiKeyIsValid = false;
-		this.apiKey = null;
 	}
 
 	/**
